@@ -15,11 +15,11 @@
         }
     ],
     "require": {
-        "php":       ">=5.3.0",
-        "twig/twig": "~1.14"
+        "php":       ">=7.2",
+        "twig/twig": "~1.2"
     },
     "require-dev": {
-	"phpunit/phpunit": "~3.7"
+	"phpunit/phpunit": "~8.0"
     },
     "autoload": {
         "psr-0": { "Mandango\\Mondator\\": "src/" }
