@@ -14,7 +14,7 @@ namespace Mandango\Mondator\Tests;
 use Mandango\Mondator\Definition\Definition;
 use Mandango\Mondator\Dumper;
 
-class DumperTest extends \PHPUnit_Framework_TestCase
+class DumperTest extends \PHPUnit\Framework\TestCase
 {
     public function testConstructor()
     {
