@@ -13,7 +13,7 @@ namespace Mandango\Mondator\Tests;
 
 use Mandango\Mondator\Output;
 
-class OutputTest extends \PHPUnit_Framework_TestCase
+class OutputTest extends \PHPUnit\Framework\TestCase
 {
     public function testConstructor()
     {
